Move graph demo to notebook.
